Single request class
