--- modules/transport-netty4/src/main/java/org/opensearch/http/netty4/AbstractNetty4HttpRequest.java
@@ -1,142 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.http.netty4;
-
-import org.opensearch.rest.RestRequest;
-
-import java.util.AbstractMap;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import io.netty.handler.codec.http.HttpHeaders;
-import io.netty.handler.codec.http.HttpMethod;
-import io.netty.handler.codec.http.HttpRequest;
-
-public abstract class AbstractNetty4HttpRequest {
-
-    protected HttpHeadersMap headers;
-    protected Exception inboundException;
-
-    protected RestRequest.Method getHttpMethod(HttpRequest request) {
-        HttpMethod httpMethod = request.method();
-        if (httpMethod == HttpMethod.GET) return RestRequest.Method.GET;
-
-        if (httpMethod == HttpMethod.POST) return RestRequest.Method.POST;
-
-        if (httpMethod == HttpMethod.PUT) return RestRequest.Method.PUT;
-
-        if (httpMethod == HttpMethod.DELETE) return RestRequest.Method.DELETE;
-
-        if (httpMethod == HttpMethod.HEAD) {
-            return RestRequest.Method.HEAD;
-        }
-
-        if (httpMethod == HttpMethod.OPTIONS) {
-            return RestRequest.Method.OPTIONS;
-        }
-
-        if (httpMethod == HttpMethod.PATCH) {
-            return RestRequest.Method.PATCH;
-        }
-
-        if (httpMethod == HttpMethod.TRACE) {
-            return RestRequest.Method.TRACE;
-        }
-
-        if (httpMethod == HttpMethod.CONNECT) {
-            return RestRequest.Method.CONNECT;
-        }
-
-        throw new IllegalArgumentException("Unexpected http method: " + httpMethod);
-    }
-
-    /**
-     * A wrapper of {@link HttpHeaders} that implements a map to prevent copying unnecessarily. This class does not support modifications
-     * and due to the underlying implementation, it performs case insensitive lookups of key to values.
-     *
-     * It is important to note that this implementation does have some downsides in that each invocation of the
-     * {@link #values()} and {@link #entrySet()} methods will perform a copy of the values in the HttpHeaders rather than returning a
-     * view of the underlying values.
-     */
-    protected static class HttpHeadersMap implements Map<String, List<String>> {
-
-        private final HttpHeaders httpHeaders;
-
-        HttpHeadersMap(HttpHeaders httpHeaders) {
-            this.httpHeaders = httpHeaders;
-        }
-
-        @Override
-        public int size() {
-            return httpHeaders.size();
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return httpHeaders.isEmpty();
-        }
-
-        @Override
-        public boolean containsKey(Object key) {
-            return key instanceof String && httpHeaders.contains((String) key);
-        }
-
-        @Override
-        public boolean containsValue(Object value) {
-            return value instanceof List && httpHeaders.names().stream().map(httpHeaders::getAll).anyMatch(value::equals);
-        }
-
-        @Override
-        public List<String> get(Object key) {
-            return key instanceof String ? httpHeaders.getAll((String) key) : null;
-        }
-
-        @Override
-        public List<String> put(String key, List<String> value) {
-            throw new UnsupportedOperationException("modifications are not supported");
-        }
-
-        @Override
-        public List<String> remove(Object key) {
-            throw new UnsupportedOperationException("modifications are not supported");
-        }
-
-        @Override
-        public void putAll(Map<? extends String, ? extends List<String>> m) {
-            throw new UnsupportedOperationException("modifications are not supported");
-        }
-
-        @Override
-        public void clear() {
-            throw new UnsupportedOperationException("modifications are not supported");
-        }
-
-        @Override
-        public Set<String> keySet() {
-            return httpHeaders.names();
-        }
-
-        @Override
-        public Collection<List<String>> values() {
-            return httpHeaders.names().stream().map(k -> Collections.unmodifiableList(httpHeaders.getAll(k))).collect(Collectors.toList());
-        }
-
-        @Override
-        public Set<Entry<String, List<String>>> entrySet() {
-            return httpHeaders.names()
-                .stream()
-                .map(k -> new AbstractMap.SimpleImmutableEntry<>(k, httpHeaders.getAll(k)))
-                .collect(Collectors.toSet());
-        }
-    }
-}
