+++ modules/transport-netty4/src/main/java/org/opensearch/http/netty4/Netty4HttpRequest.java
@@ -32,17 +32,20 @@
 
 package org.opensearch.http.netty4;
 
+import org.opensearch.core.common.bytes.BytesArray;
 import org.opensearch.core.common.bytes.BytesReference;
 import org.opensearch.core.rest.RestStatus;
-import org.opensearch.http.HttpRequest;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.transport.netty4.Netty4Utils;
 
+import java.util.AbstractMap;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
@@ -51,50 +54,48 @@ import io.netty.handler.codec.http.DefaultHttpHeaders;
 import io.netty.handler.codec.http.FullHttpRequest;
 import io.netty.handler.codec.http.HttpHeaderNames;
 import io.netty.handler.codec.http.HttpHeaders;
+import io.netty.handler.codec.http.HttpMethod;
+import io.netty.handler.codec.http.HttpRequest;
 import io.netty.handler.codec.http.cookie.Cookie;
 import io.netty.handler.codec.http.cookie.ServerCookieDecoder;
 import io.netty.handler.codec.http.cookie.ServerCookieEncoder;
 
-public class Netty4HttpRequest extends AbstractNetty4HttpRequest implements HttpRequest {
+public class Netty4HttpRequest implements org.opensearch.http.HttpRequest {
 
-    private final FullHttpRequest request;
+    private final HttpRequest request;
     private final BytesReference content;
+    private final HttpHeadersMap headers;
     private final AtomicBoolean released;
+    private final Exception inboundException;
     private final boolean pooled;
 
-    Netty4HttpRequest(FullHttpRequest request) {
+    public Netty4HttpRequest(HttpRequest request) {
         this(
             request,
             new HttpHeadersMap(request.headers()),
             new AtomicBoolean(false),
             true,
-            Netty4Utils.toBytesReference(request.content())
+            (request instanceof FullHttpRequest) ? Netty4Utils.toBytesReference(((FullHttpRequest) request).content()) : BytesArray.EMPTY
         );
     }
 
-    Netty4HttpRequest(FullHttpRequest request, Exception inboundException) {
+    Netty4HttpRequest(HttpRequest request, Exception inboundException) {
         this(
             request,
             new HttpHeadersMap(request.headers()),
             new AtomicBoolean(false),
             true,
-            Netty4Utils.toBytesReference(request.content()),
+            (request instanceof FullHttpRequest) ? Netty4Utils.toBytesReference(((FullHttpRequest) request).content()) : BytesArray.EMPTY,
             inboundException
         );
     }
 
-    private Netty4HttpRequest(
-        FullHttpRequest request,
-        HttpHeadersMap headers,
-        AtomicBoolean released,
-        boolean pooled,
-        BytesReference content
-    ) {
+    private Netty4HttpRequest(HttpRequest request, HttpHeadersMap headers, AtomicBoolean released, boolean pooled, BytesReference content) {
         this(request, headers, released, pooled, content, null);
     }
 
     private Netty4HttpRequest(
-        FullHttpRequest request,
+        HttpRequest request,
         HttpHeadersMap headers,
         AtomicBoolean released,
         boolean pooled,
@@ -111,7 +112,36 @@ public class Netty4HttpRequest extends AbstractNetty4HttpRequest implements Http
 
     @Override
     public RestRequest.Method method() {
-        return getHttpMethod(request);
+        HttpMethod httpMethod = request.method();
+        if (httpMethod == HttpMethod.GET) return RestRequest.Method.GET;
+
+        if (httpMethod == HttpMethod.POST) return RestRequest.Method.POST;
+
+        if (httpMethod == HttpMethod.PUT) return RestRequest.Method.PUT;
+
+        if (httpMethod == HttpMethod.DELETE) return RestRequest.Method.DELETE;
+
+        if (httpMethod == HttpMethod.HEAD) {
+            return RestRequest.Method.HEAD;
+        }
+
+        if (httpMethod == HttpMethod.OPTIONS) {
+            return RestRequest.Method.OPTIONS;
+        }
+
+        if (httpMethod == HttpMethod.PATCH) {
+            return RestRequest.Method.PATCH;
+        }
+
+        if (httpMethod == HttpMethod.TRACE) {
+            return RestRequest.Method.TRACE;
+        }
+
+        if (httpMethod == HttpMethod.CONNECT) {
+            return RestRequest.Method.CONNECT;
+        }
+
+        throw new IllegalArgumentException("Unexpected http method: " + httpMethod);
     }
 
     @Override
@@ -127,27 +157,32 @@ public class Netty4HttpRequest extends AbstractNetty4HttpRequest implements Http
 
     @Override
     public void release() {
+        assert request instanceof FullHttpRequest : "release can only be called when underlying request object is of type FullHttpRequest";
         if (pooled && released.compareAndSet(false, true)) {
-            request.release();
+            FullHttpRequest req = (FullHttpRequest) request;
+            req.release();
         }
     }
 
     @Override
-    public HttpRequest releaseAndCopy() {
+    public org.opensearch.http.HttpRequest releaseAndCopy() {
+        assert request instanceof FullHttpRequest
+            : "releaseAndCopy can only be called when underlying request object is of type FullHttpRequest";
         assert released.get() == false;
         if (pooled == false) {
             return this;
         }
+        FullHttpRequest req = (FullHttpRequest) request;
         try {
-            final ByteBuf copiedContent = Unpooled.copiedBuffer(request.content());
+            final ByteBuf copiedContent = Unpooled.copiedBuffer(req.content());
             return new Netty4HttpRequest(
                 new DefaultFullHttpRequest(
-                    request.protocolVersion(),
-                    request.method(),
-                    request.uri(),
+                    req.protocolVersion(),
+                    req.method(),
+                    req.uri(),
                     copiedContent,
-                    request.headers(),
-                    request.trailingHeaders()
+                    req.headers(),
+                    req.trailingHeaders()
                 ),
                 headers,
                 new AtomicBoolean(false),
@@ -179,27 +214,29 @@ public class Netty4HttpRequest extends AbstractNetty4HttpRequest implements Http
     @Override
     public HttpVersion protocolVersion() {
         if (request.protocolVersion().equals(io.netty.handler.codec.http.HttpVersion.HTTP_1_0)) {
-            return HttpRequest.HttpVersion.HTTP_1_0;
+            return org.opensearch.http.HttpRequest.HttpVersion.HTTP_1_0;
         } else if (request.protocolVersion().equals(io.netty.handler.codec.http.HttpVersion.HTTP_1_1)) {
-            return HttpRequest.HttpVersion.HTTP_1_1;
+            return org.opensearch.http.HttpRequest.HttpVersion.HTTP_1_1;
         } else {
             throw new IllegalArgumentException("Unexpected http protocol version: " + request.protocolVersion());
         }
     }
 
     @Override
-    public HttpRequest removeHeader(String header) {
+    public org.opensearch.http.HttpRequest removeHeader(String header) {
         HttpHeaders headersWithoutContentTypeHeader = new DefaultHttpHeaders();
         headersWithoutContentTypeHeader.add(request.headers());
         headersWithoutContentTypeHeader.remove(header);
         HttpHeaders trailingHeaders = new DefaultHttpHeaders();
-        trailingHeaders.add(request.trailingHeaders());
-        trailingHeaders.remove(header);
+        if (request instanceof FullHttpRequest) {
+            trailingHeaders.add(((FullHttpRequest) request).trailingHeaders());
+            trailingHeaders.remove(header);
+        }
         FullHttpRequest requestWithoutHeader = new DefaultFullHttpRequest(
             request.protocolVersion(),
             request.method(),
             request.uri(),
-            request.content(),
+            (request instanceof FullHttpRequest) ? ((FullHttpRequest) request).content() : Unpooled.EMPTY_BUFFER,
             headersWithoutContentTypeHeader,
             trailingHeaders
         );
@@ -216,7 +253,87 @@ public class Netty4HttpRequest extends AbstractNetty4HttpRequest implements Http
         return inboundException;
     }
 
-    public FullHttpRequest nettyRequest() {
+    public HttpRequest nettyRequest() {
         return request;
+    }
+
+    /**
+     * A wrapper of {@link HttpHeaders} that implements a map to prevent copying unnecessarily. This class does not support modifications
+     * and due to the underlying implementation, it performs case insensitive lookups of key to values.
+     *
+     * It is important to note that this implementation does have some downsides in that each invocation of the
+     * {@link #values()} and {@link #entrySet()} methods will perform a copy of the values in the HttpHeaders rather than returning a
+     * view of the underlying values.
+     */
+    private static class HttpHeadersMap implements Map<String, List<String>> {
+
+        private final HttpHeaders httpHeaders;
+
+        private HttpHeadersMap(HttpHeaders httpHeaders) {
+            this.httpHeaders = httpHeaders;
+        }
+
+        @Override
+        public int size() {
+            return httpHeaders.size();
+        }
+
+        @Override
+        public boolean isEmpty() {
+            return httpHeaders.isEmpty();
+        }
+
+        @Override
+        public boolean containsKey(Object key) {
+            return key instanceof String && httpHeaders.contains((String) key);
+        }
+
+        @Override
+        public boolean containsValue(Object value) {
+            return value instanceof List && httpHeaders.names().stream().map(httpHeaders::getAll).anyMatch(value::equals);
+        }
+
+        @Override
+        public List<String> get(Object key) {
+            return key instanceof String ? httpHeaders.getAll((String) key) : null;
+        }
+
+        @Override
+        public List<String> put(String key, List<String> value) {
+            throw new UnsupportedOperationException("modifications are not supported");
+        }
+
+        @Override
+        public List<String> remove(Object key) {
+            throw new UnsupportedOperationException("modifications are not supported");
+        }
+
+        @Override
+        public void putAll(Map<? extends String, ? extends List<String>> m) {
+            throw new UnsupportedOperationException("modifications are not supported");
+        }
+
+        @Override
+        public void clear() {
+            throw new UnsupportedOperationException("modifications are not supported");
+        }
+
+        @Override
+        public Set<String> keySet() {
+            return httpHeaders.names();
+        }
+
+        @Override
+        public Collection<List<String>> values() {
+            return httpHeaders.names().stream().map(k -> Collections.unmodifiableList(httpHeaders.getAll(k))).collect(Collectors.toList());
+        }
+
+        @Override
+        public Set<Entry<String, List<String>>> entrySet() {
+            return httpHeaders.names()
+                .stream()
+                .map(k -> new AbstractMap.SimpleImmutableEntry<>(k, httpHeaders.getAll(k)))
+                .collect(Collectors.toSet());
+        }
     }
 }
