+++ modules/transport-netty4/src/main/java/org/opensearch/http/netty4/Netty4HttpResponse.java
@@ -46,7 +46,7 @@ public class Netty4HttpResponse extends DefaultFullHttpResponse implements HttpR
 
     private final HttpHeaders requestHeaders;
 
-    public Netty4HttpResponse(HttpHeaders requestHeaders, HttpVersion version, RestStatus status, BytesReference content) {
+    Netty4HttpResponse(HttpHeaders requestHeaders, HttpVersion version, RestStatus status, BytesReference content) {
         super(version, HttpResponseStatus.valueOf(status.getStatus()), Netty4Utils.toByteBuf(content));
         this.requestHeaders = requestHeaders;
     }
